complete rewrite: first working draft!
now gochan-select* supports sending to (or receiving from) muliple
channels too. now we should be more or less on par with the go
language's API.

nothing from the original gochan.scm (except info and the go syntax
and a couple other things) is reused, really. enabling gochan-select
to support sending as well as receiving, this introduced some serious
rethinking.

the semaphore-concept from gochan1 was kept, but slightly extended and
perhaps simplified. recv and send are special cases of gochan-select,
as this code now reflects.

gochan.scm went from 392 to 195 lines of code which is also nice, even
though we're missing some features:

- buffers
- timeouts

perhaps these will be easier to write now. note that "receive" was
renamed to "recv" because it's just as long as "send" which is nice.