feat: adds gochan-close
so now you can close your channels. all recievers will be notified,
just like in go, with the `ok` flag set to #f.

note the addition of the `ok` flag set in a send-clause of
gochan-select! this means you can do this:

(gochan-select ((chan <- 123 ok?) (if ok? (print "chan closed!!"))))

it's possible this is be better than a panic (like golang does it),
and it's also consistent with how receive on a closed channel behaves.