api: don't execute gochan-select body unless ok by default
if we simply ignore the ok flag and execute the body regardless of
what it is, this quickly becomes a pattern:

  (let loop ()
     (gochan-select
       ((chan -> msg ok)
          (when ok
             (do-stuff msg)
             (loop)))))

and there isn't really any way around this, we always need to check
"ok". if we want to make "ok" an optional binding, however, we should
default to _not_ executing the body of the clause (and just returning
void) when "ok" is #f. this shortens the code above my removing the
need for (when ok ...).

if you actually _do_ want to execute the body regardless of what "ok"
is, just include ok in the variable binding. this should improve the
API.