api major: invert "ok flag" to "fail flag"
this is to make it more useful:

 - an ok-flag is always either #f (closed)/#t (non-closed)
 - a fail-flag is #f for no failure, and can carry data in it when its closed

this allows us to use closed channels as "broadcasting channels",
broadcasting their fail-flag to all listeners and any newcomers.

fail-flag should probably be called the "close flag", because
currently the only reason a send/recv can fail is that it is closed.

note that this patch mostly inverts #t to #f and does some renaming in
gochan.scm.