bugfix critical: adds missing mutex-unlock!
I can't believe this has been there unnoticed for so long. The bug
occurs if you:

thread 1 does

    (gochan-select ((c -> _)) (else #t))

thread 2 does

    (gochan-select ((c <- '_)))

and the scheduler decides to time it so that:

thread 2 runs when the semaphore of thread 1 has been registered in
the gochan. I don't know how to make a test-case for that!